+++ Tests/TuistCoreTests/Models/ProductTests .swift
@@ -51,6 +51,7 @@ final class ProductTests: XCTestCase {
         let got = Product.forPlatform(.macOS)
         let expected: [Product] = [
             .app,
+            .commandLineTool,
             .staticLibrary,
             .dynamicLibrary,
             .framework,
@@ -76,7 +77,7 @@ final class ProductTests: XCTestCase {
 
     func test_runnable() {
         Product.allCases.forEach { product in
-            if [.app, .appClip].contains(product) {
+            if [.app, .appClip, .commandLineTool].contains(product) {
                 XCTAssertTrue(product.runnable)
             } else {
                 XCTAssertFalse(product.runnable)