+++ fixtures/command_line_tool_with_static_dependencies/CommandLineTool/main.swift
@@ -1,0 +1,4 @@
+import Foundation
+import StaticLib
+
+let staticClass = StaticLib()